styling for header and footer fixed
