+++ custom-app/src/main.tsx
@@ -1,14 +1,14 @@
 import './main.css'
-
-import { NavFooter, NavHeader } from './NavHeader'
+import { NavHeader } from './NavHeader'
+import { NavFooter } from './NavFooter'
 import { onDOMContentMutate, renderInContainer } from './utils'
 
 function render() {
   // render custom header
-  renderInContainer(NavHeader, document.getElementById('fern-header'))
+  renderInContainer(NavHeader, document.getElementById('fern-header'));
 
   // render custom footer
-  renderInContainer(NavFooter, document.getElementById('fern-footer'))
+  renderInContainer(NavFooter, document.getElementById('fern-footer'));
 }
 
 onDOMContentMutate(render)
